+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Rename `RepliconClientStatus` to `ClientState` and `RepliconServerStatus` to `ServerState`. They are now regular Bevy states. As result, we now require `StatesPlugin` to be added. It's present by default in `DefaultPlugins`, but with `MinimalPlugins` you have to add it manually.
-- Replace `TickPolicy` with `TickSchedule` and `ServerPlugin::tick_policy` with `ServerPlugin::tick_schedule`. Instead of default `TickPolicy::MaxTickRate`, it now `TickSchedule::FixedPostUpdate`. Insert `Time<Fixed>` to control how often it runs.
+- Replace `ServerPlugin::tick_policy` with `ServerPlugin::tick_schedule`, which specifies the schedule where the tick increments. By default, it's `FixedPostUpdate`.
 - Make custom entity ser/de compatible with `serde` attributes.
 - All contexts now store `AppTypeRegistry` instead of `TypeRegistry`. To get `TypeRegistry`, call `AppTypeRegistry::read`.
 - All events now use `ClientId` wrapper instead of `Entity`.
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replace `ReplicationBundle` with `BundleRules`, which returns only `Vec<ComponentRule>` and uses an associated constant to customize the priority.
 - Hide `ServerEntityMap` mutation methods from public API.
 - Hide `BufferedMutations` from public API.
+- Hide `server::increment_tick` from public API.
 - `ExampleClient::new` now accepts `impl Into<SocketAddr>`. This makes it easier for backend developers to port examples since they usually specify IP address.
 
 ## Removed
@@ -48,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `entity_serde::serialize_entity` and `entity_serde::deserialize_entity`. Use `postcard_utils::entity_to_extend_mut` and `postcard_utils::entity_from_buf` respectively; just swap the argument order.
 - `SERVER`. Use `ClientId::Server` instead.
 - `ReplicationMode::Periodic`. Use `PriorityMap` instead.
+- `TickPolicy`. You now can set the schedule directly or configure how often it runs by inserting `Time::<Fixed>`.
 - All provided run conditions. Just use `in_state` or `OnEnter`/`OnExit` with `ServerState` and `ClientState` instead. `server_or_singleplayer` is just `in_state(ClientState::Disconnected)`.
 
 ## [0.34.4] - 2025-07-29